+++ EEGAnalysis/src/filters/ShortTimeFourier.java
@@ -5,7 +5,7 @@ import filters.utils.ChunkedData;
 import filters.utils.Filter;
 
 public class ShortTimeFourier extends Filter {
-	
+
 	// @TODO
 	public static double[][] compute(double[][] data) {
 		int fs = 100;
@@ -14,8 +14,8 @@ public class ShortTimeFourier extends Filter {
 		int freqLowerLimit = (int) (2 * resolutionFactor);
 		int freqUpperLimit = (timeResolution <= fs) ? timeResolution/2 : fs/2;
 		ChunkedData chunked = new ChunkedData(data[Y], fs, timeResolution, 0);
-		int yFqTo = (int) ((timeResolution / (double)fs) * (double)freqUpperLimit);
-		int yFqFrom = (int) ((timeResolution / (double)fs) * (double)freqLowerLimit);
+		int yFqTo = (int) ((timeResolution / (double)fs) * freqUpperLimit);
+		int yFqFrom = (int) ((timeResolution / (double)fs) * freqLowerLimit);
 		int yLen = yFqTo - yFqFrom;
 		Logger.log("request array X len:" + chunked.getNumberOfChunk());
 		Logger.log("request array Y len:" + yLen);
@@ -29,13 +29,13 @@ public class ShortTimeFourier extends Filter {
 		for(int i=0; i < chunked.getNumberOfChunk(); i++) {
 			stf[X][i] = i;
 		}
-		
+
 		double[] psd = null;
 		while(chunked.hasNextChunk()) {
 			psd = EnergySpectralDensity.compute(chunked.getChunk(), fs);
 			try {
 				for(int i=yFqFrom; i<yFqTo; i++) {
-					stf[Z][(chunked.getChunkPosition()*yLen) + i - yFqFrom] = Math.log(psd[i]); 
+					stf[Z][(chunked.getChunkPosition()*yLen) + i - yFqFrom] = Math.log(psd[i]);
 				}
 			} catch (Exception e) {
 				Logger.log("	yFqFrom==" + yFqFrom);
@@ -50,7 +50,7 @@ public class ShortTimeFourier extends Filter {
 		for(int i=yFqFrom; i<yFqTo; i++) {
 			stf[Y][i-yFqFrom] = i;
 		}
-		
+
 		return stf;
 	}
 }
